move gyroradius tests into test_lengths.py
